MLOps/reranker_cloud_demo/sagemaker_pipeline/requirements-sagemaker.txt
uvicorn==0.24.0
sentence-transformers==4.1.0
streamlit==1.45.1
torch==2.3.1
transformers==4.42.4
datasets==2.20.0
pandas==2.2.2
numpy==1.26.4
pydantic==2.5.0
python-multipart==0.0.6
psutil==5.9.0
supabase==2.16.0
mlflow
boto3
python-dotenv
